(PC-20106) refactor(notifications): make features/notifications following the code structure
